Fix syslog test failure about fixture tbinfo (#13821)
Description of PR
Syslog test failed for complaining about tbinfo usage as it's a fixture. Solution is to add tbinfo as fixture in the orch_logrotate_setup. Then referencing tbinfo will be the yield of tbinfo which is a dict

What is the motivation for this PR?
Fix the test failure for syslog test_logrotate.

How did you do it?
How did you verify/test it?
Tested on OC test bed and see test pass now.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>