Update some references of ceos.yml to the new path (#14366)
What is the motivation for this PR?
In PR #14200, it updated the file path of ceos.yml. But in our code, there still exists some references to the file which are not be updated to the new path. In this PR, we update these references to the new path.

How did you do it?
Update some references of ceos.yml to the new path.

Signed-off-by: yenlu <yenlu@cisco.com>