[Chassis] Recover chassis by config reload all RP/LCs (#14368)
Description of PR
Summary:

Extend the wait time of the readiness check for chassis devices
Provide a modular-chassis specific recover way, for example, if the iBGP on the LC can't be up, do config reload on the faulty LC doesn't help, the required action is to config reload on the RP. Hence for the chassis device, directly config reload all the RP/LCs

Approach
What is the motivation for this PR?
The current recover doesn't help chassis testbed well when it's in the bad state.

How did you do it?
Extend the wait time of the readiness check for chassis devices
Provide a modular-chassis specific recover way, config reload all the devices.
How did you verify/test it?
In the nightly test and PR tests

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>