[wol_test] Test expected failure like invalid parameters, down interface etc. (#14387)
What is the motivation for this PR?
Wol has the ability to filter invalid parameters, we need to test this ability.

How did you do it?
Run wol with invalid parameters and check the output and return code of command.

How did you verify/test it?
Run on testbed

Any platform specific information?
Mx topo only.

Signed-off-by: yenlu <yenlu@cisco.com>