Upstream arp/test_stress_arp.py changes to 202405 branch (#14420)
There are 3 issues with arp/test_stress_arp.py:

It is doing a redundant subtraction.
In both variants, it calculates space available as (available - used)
instead of (total size - used), giving an incorrect value for available
space.

Not enough allowance/wiggle room for hash collision. In both variants,
it expects only 100 hash collisions which is very little. For specific
SKUs, there may be a further performance impact as some hash functions
is not as performant for incremental values, which is what is used for
this test.

The test will crash if cleanup fails.
Sometimes arp cache cleanup will fail if it is done too often. Allowing
the test to continue after the failure as the failure is unrelated to
the test and being unable to clean implies it is clean.

This patch fixes these 3 issues by:

- Removing the redundant subtraction.
- Change hash collision allowance from 100 to 250.
- Add a catch for failing clear_dut_arp_cache().