Remove dependence in bfd_helpers.py (#14431)
What is the motivation for this PR?
In the script bfd_helpers.py, there was an import from the platform_tests folder. To avoid importing from other folders within our code, this PR removes that dependency.

How did you do it?
Rewrite the function in the script bfd_helpers.py and remove dependency.

Signed-off-by: yenlu <yenlu@cisco.com>