Make oscillation fix persistent across reboots and config reload (#14089)
What is the motivation for this PR?
#13004, introduces a fix to disable oscillations at the session level for dualtor.

The fix however is not persistent across reboots and config reload.

During the sanity checks that run in the beginning of the tests, https://github.com/sonic-net/sonic-mgmt/blob/master/tests/conftest.py#L77, if it is found that any interfaces or BGP sessions is/are down (which could be due to a test that run before the current test and contaminated the testbed), a config_reload and then further a reboot is issued, which clears off this config to disable oscillation, resulting in a flakiness when the test runs or a sanity check failure again.

How did you do it?
The proposed fix is to do a config save -y of this config so as to make it persistent across reboots and config reload.

How did you verify/test it?
Ran the tests that were flaky due to oscillations.
a) acl/test_acl.py
b) drop_packets/drop_packets.py
c) fib/test_fib.py

Any platform specific information?

Signed-off-by: yenlu <yenlu@cisco.com>