Add test_nbr_health.py into PR test (#14473)
What is the motivation for this PR?
We consider the script test_nbr_health.py as applicable to all topologies. In this PR, we update the topology mark for the script to enable it to run on all topologies and add it to the onboarding t0 and t1 PR checkers.

How did you do it?
Update the topology mark for script test_nbr_health.py to enable it to run on all topologies and add it to the onboarding t0 and t1 PR checkers.