Multi-asic support for snmp_queue_counters test (#13637)
Approach
What is the motivation for this PR?
Currently, 'test_snmp_queue_counters' test doesn't have multi-asic support and the test is failing for multi-asic chassis due to the lack of support.
How did you do it?
Added multi-asic support for test 'test_snmp_queue_counters' and made sure the current test logic is not affected.
How did you verify/test it?
Ran the above-mentioned test case on a T2 chassis and verified.

Signed-off-by: yenlu <yenlu@cisco.com>