tests-common: Handle SSHConsoleConn being closed prematurely (#14511)
What is the motivation for this PR?
dut_console/test_idle_timeout was failing due to the console port being blocked (on a specific device with a blocked port), but the error was a generic OSError.

How did you do it?
Added a more descriptive error for ease of diagnosing and resolving the issue.

How did you verify/test it?
Ran on the affected testbed, resulting in the new error being thrown when the device's console port is blocked.

Any platform specific information?
There may be platforms which do not use the same Port is in use. Closing connection... message - however if such a case is encountered, it would be trivial to enhance the check to account for platform differences.

Documentation
This is a minor enhancement, so I have not updated any documentation.

Signed-off-by: yenlu <yenlu@cisco.com>