[Bugfix] Bugfix for getting maximum matching length of conditional mark.  (#14515)
Description of PR
We changed the matching rule of conditional mark in PR #14395. There is a bug when getting the maximum matching length. In this PR, we fix this bug.

Approach
What is the motivation for this PR?
We changed the matching rule of conditional mark in PR #14395. There is a bug when getting the maximum matching length. In this PR, we fix this bug.

co-authorized by: jianquanye@microsoft.com