Relocate `gu_utils.py` to `tests/common` for dependency reduction (#14476)
What is the motivation for this PR?
Several scripts import the gu_utils.py module from the generic_config_updater folder. To minimize cross-module dependencies, we move this module to the tests/common folder.

How did you do it?
Move the module gu_utils.py to the tests/common folder

Signed-off-by: yenlu <yenlu@cisco.com>