[pfcwd] Add half of polling time as compensation for test_pfcwd_timer_accuracy (#14589)
What is the motivation for this PR?
test_pfcwd_timer_accuracy case is flaky on Arisa platform. Sometimes the detect time is larger than the config detect time.
Both the config detect time and polling time are 400ms, and most of the real detect times range between 800 ~ 1000 ms.
Based on lua script log, in the failure loop (the detect time is larger than the config detect time), it took 3 polling durations to trigger the pfc storm, and in most of these cases, there was a little traffic in the first loop pooling duration. Suppose the timestamp for the script to send PFC frames was at the end of the first polling duration. then cause there were no enough PFC received and trigger the pfc storm in the third polling loop.

How did you do it?
Add half of polling time as compensation for the detect config time.

How did you verify/test it?
Run the case

Signed-off-by: yenlu <yenlu@cisco.com>