[Snappi]: Support to clear and check fabric counters (#14097)
Description of PR
Currently, there is no mechanism in the Snappi testcases to check for the fabric counters.

Purpose of this PR is to have support to clear the counters before test-case execution. The dut host can be supervisor card or line cards.

Second function will check for the CRC and uncorrectable FEC errors. The check should assert if the count is nonzero.

Summary:
Fixes # (issue)
#14095

Approach
What is the motivation for this PR?
Currently there is no mechanism to verify the fabric counters on the DUT. The function is added to check for the CRC and uncorrectable errors on line card/s and supervisor.

How did you do it?
Two simple functions are added. First to clear the counter and second to parse for errors. If the errors are non-zero, the test will ASSERT.

How did you verify/test it?
Tested on local clone.

co-authorized by: jianquanye@microsoft.com