[COPP][720DT] Fix copp test failure for 720dt (#14627)
What is the motivation for this PR?
COPP rate of queue4_group3 for Mgmt devices is set to 300. But in copp_test it's expected to be 100 for Arista 720DT. This PR is to fix that.
https://github.com/sonic-net/sonic-buildimage/blob/2c47e35472b1f9fd5801a884ce7993d873c07afa/files/image_config/copp/copp_cfg.j2#L37

How did you do it?
Set expected copp rate for 720dt

How did you verify/test it?
Run test_copp.py in m0/t0/t1 topos, all passed

Signed-off-by: yenlu <yenlu@cisco.com>