Merged PR 7307: chore: add completeness param for bfd
Currently, we are having a large iteration number for `bfd/test_bfd_static_route.py::test_bfd_flap`, causing long running time in Nightly. To reduce the time, we are going to start with the debug completeness level (i.e. one iteration) and increase it in the future.

----
#### AI description  (iteration 1)
#### PR Classification
Chore: Add completeness parameter for BFD.

#### PR Summary
This pull request adds a new parameter to the pipeline configuration for BFD testing.
- Changes in `/.azure-pipelines/elastictest/cisco/8800.t2.202405.yml` to include `--completeness_level=debug` for `bfd/test_bfd_static_route.py`.