Merged PR 7346: Cleanup devices moved to TK5 in 2023-11
Cleanup devices moved to TK5 in 2023-11

We moved lots of devices from starlab B43 to TK5 during 2024-11.

Many of the devices are still in old connection graph files and inventory files.

This change is to cleanup these information to avoid confusion.

Device strtk5-7260-fanout-22 has been renamed to strtk5-7260-7. The device name was not updated in the tk5 PDU links csv file. This change also fixed that.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
This pull request removes references to devices that have been moved to TK5 as of November 2023.
- `sonic_str_pdu_links.csv`: Removed multiple PDU entries for devices moved to TK5.
- `str`: Deleted several device entries and their associated host configurations.
- `sonic_str_devices.csv`: Removed device entries for devices moved to TK5.
- `sonic_str_console_links.csv`: Deleted console link entries for devices moved to TK5.
- `sonic_str_links.csv`: Removed network link entries for devices moved to TK5.