[Bugfix] Fix NameError imported by PR #14672 (#14723)
What is the motivation for this PR?
In PR #14672, we moved some functions from platform_tests/verify_dut_health.py to common/platform/device_utils.py, but forgot to move the defination of variable test_report, which will cause an NameError. In this PR, we define this variable in common/platform/device_utils.py to fix this issue.

How did you do it?
Define the variable in common/platform/device_utils.py to fix this issue.

Signed-off-by: yenlu <yenlu@cisco.com>