Enable fdb and qos tests for t0-standalone-* topology (#14733)
What is the motivation for this PR?
Solve 202311 branch conflicts in the PR #13019 .The t0-standalone topologies should support the FDB and QoS test cases, however we have per topology filters on these tests, so these tests will be skipped or failed.

How did you do it?
Manually cherry-picked it and solved conflicts. This change updates the whitelists that we have to allow the tests to be run.

How did you verify/test it?
Run locally with the new settings, both fdb and qos tests are not filtered out anymore.