Add new test for vlan subnet decap (#14720)
What is the motivation for this PR?
In Azure network, Netscan uses IP-decap to check for route blackholes, but servers do not support decapsulation, making it impossible to perform server-to-server checks. To address this, we developed the subnet decap feature to detect if the link between the server and T0 is functioning properly. Additionally, new test cases are required to ensure the functionality of subnet decap is working as expected.

How did you do it?
Add new test to verify IPinIP packet targeting at VLAN subnet is decapsulated (T0 only)
Ref test gap issue: #13120
Ref test plan: #12990

How did you verify/test it?
Test on broadcom td3 asic and KVM