[Bugfix] Restoring wan-pub topology mapping in test scripts validation (#14758)
What is the motivation for this PR?
In PR #14437, we removed the mapping for the wan-pub topology name during test script validation, as there is no WAN checker in our PR test. However, this removal resulted in skipped scripts for that topology not being collected. In this PR, we are restoring the mapping to address this issue.

How did you do it?
Restore the mapping.

Signed-off-by: yenlu <yenlu@cisco.com>