Fix Python3 str decode issues (#14332)
What is the motivation for this PR?
This fixes an issue due to Python 3 migration, where decode() is no longer a method of str.

How did you do it?
For cases where we actually need the unicode-escape conversion, I convert to bytes first, then re-decode to str to decode the escape sequences. In other cases where we are just taking the raw output of a command, I just removed the decode().

How did you verify/test it?
I ran the relevant sonic-mgmt tests and verified that the tests now pass and no longer raise exceptions due to decode() not being a method of str or AnsibleUnsafeText.

Signed-off-by: yenlu <yenlu@cisco.com>