Generate console connection information for devices (#14523)
What is the motivation for this PR?
For devices, generate console connection entries if this information is available. This allows people to connect to those devices via console if regular SSH access is not working.

Also, place arguments to the script into multiple groups, for readability purposes, and make the SSH config syntax checker be validated at argument validation time instead of later.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>