tacacs/test_ro_disk.py: Increase wait time for modular chassis after reboot (#14718)
Description of PR
The testcase only waits for 5 mins after Supervisor reboot, this is not enough in case of modular chassis. This causes all testcases to fail after this test stating that Host is unavailable.

Summary:
Fixes # (issue)

Approach
What is the motivation for this PR?
Wait time of 5 minutes is not enough for supervisor node. Especially in this testcase where ssh check is disabled.

How did you do it?
Increased the time for supervisor node.

How did you verify/test it?
Validated on modular chassis with T2 profile

co-authorized by: jianquanye@microsoft.com