Always run postboot_setup in runRebootTest (#14347)
What is the motivation for this PR?
In event of an exception in runRebootTest, postboot_setup may not run.

How did you do it?
Move the call to the finally block to ensure it is always run.

How did you verify/test it?
Verified running test_upgrade_path on Arista device.

Signed-off-by: yenlu <yenlu@cisco.com>