Merged PR 7511: chore: adjust date to start 00:00
Without specifying the time, running `show techsupport --since="x days ago"` it will create a `/tmp/reference` file with the `current time + x days ago`.

This will lead to us missing some of the logs on the same day if the time of the log was earlier. Changing this will make sure that we consider logs from 00:00

Signed-off-by: Austin Pham <austinpham@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
Chore to adjust date format in pipeline configuration.

#### PR Summary
This pull request updates the date format in the pipeline configuration to start at 00:00.
- Changes in `/.azure-pipelines/elastictest/cisco/8800.t2.202405.yml` to adjust `--posttest_show_tech_since` parameter to `-5days00:00`.