Revert "Add ASIC DB validation to ACL tests (#13574)" (#14565)
Description of PR
SAI validation work is in progress. The new implementation and approach are different and hence reverts commit b706c4e.

Approach
What is the motivation for this PR?
SAI validation needs to happen at the setup / config step of the tests and not when tests run. Hence reverting.

How did you do it?
Reverted the commit.