[Snappi] multidut test case changes to support T0/T1 testbed. (#14742)
Description of PR
This PR is to accommodate the changes in the test cases to support T0/T1 testbeds after the infra change in #14361.

Summary:
Test case changes for #14220 and #14144

Approach
What is the motivation for this PR?
Consolidate T0/T1 and T2 test cases.

How did you do it?
Main change:
- Change the test case to use all available ports for T0/T1 test cases.
- Same change is repeated in multidut PFC/PFCWD/ECN test cases.

Minor change:
- fix the following error:

>       lossless_prio = int(lossless_prio)
E       TypeError: int() argument must be a string, a bytes-like object or a number, not 'list'
How did you verify/test it?
Ran it on T0/T1 and T2 testbed. failures are seen but not related to this change.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>