Bugfix (#14878)
Description of PR
This PR addresses a bug introduced in PR #14551. In that PR, we refactored several fixtures, but they were not imported in the test_mgmt_ipv6_only.py script, leading to errors.

Approach
What is the motivation for this PR?
This PR addresses a bug introduced in PR #14551. In that PR, we refactored several fixtures, but they were not imported in the test_mgmt_ipv6_only.py script, leading to errors.

co-authorized by: jianquanye@microsoft.com