sonic-mgmt: improve asic identification
Device ASIC identification is achieved by
whole line matches from the output of lspci,
which is excessive and subject to fail due to
unforeseeable changes in such output.

This change reduces the string matching to
specific unique differentiators in the output
from lspci, while also future-proofing against
similar changes in the lspci that could
foreseeably occur.