Skipping EverflowV4 EgressAcl EgressMirror for Cisco-8000 (#14964)
Description of PR
Skipping EverflowV4 EgressAcl EgressMirror for Cisco-8000 as the feature is currently not fully supported. Will remove the fix once the feature is validated.

Replacing Xfail with Skip for test mirror with policer since it is not supported on Cisco 8000 platforms. Xfail was causing the test to run and then xfail which is not desirable, since the testcase is not supported.

Approach
What is the motivation for this PR?
How did you do it?
How did you verify/test it?
Any platform specific information?
For Cisco Platforms

co-authorized by: jianquanye@microsoft.com