chore: change link toggle command (#14987)
Description of PR
Remove the ip netns exec asicX part from the link toggle command in BFD tests as the -n option in config interface takes care of the namespace.

Approach
What is the motivation for this PR?
We are seeing some flaky issue when toggling the interface with ip netns exec asicX in the command. Cisco suggests to remove this part from the command as the -n option already takes care of the namespace.

How did you do it?
How did you verify/test it?
I ran the updated code and can confirm it's working as expected.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>