Move test_dualtor_bgp_update_delay.py to dualtor (#14997)
What is the motivation for this PR?
We have created an issue of test script test_dualtor_bgp_update_delay.py and used conditional mark to xfail this script. So it can be moved from onboarding dualtor PR checker to dualtor PR checker. As there is no scripts in onboarding dualtor PR checker, so we comment out this checker.

Signed-off-by: yenlu <yenlu@cisco.com>