Relocate log_custom_msg to assure test results to be logged
Instead of fixture, log_custom_msg is now called in one of hook (pytest_runtest_makereport).
This is to prevent unexpected sequence of fixture teardown observed in nightly.
Calling log_custom_msg from pytest_runtest_makereport will be one of the very last hooks to be called for each test.