Fix image URLs passed in on command line not being recognized (#14943)
For start-topo-vms, an image URL can be passed in so that if the image
doesn't exist on disk, it can be downloaded. However, the type filter
for this seems to treat this type as `unicode` instead of `string`. This
could be due to some Python 2 vs Python 3 differences, but I'm not
certain on this.

Fix this by checking for both unicode and string types.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>