Adjust fixture order in arp test unknown mac (#15001)
What is the motivation for this PR?
Fix fixture order to eliminate flaky test

How did you do it?
Make the fixture flushArpFdb depend on the fixture dut_disable_arp_update, so that the fixture populateArp depends on flushArpFdb, which in turn depends on dut_disable_arp_update. This ensures that ARP updates are disabled and tables are cleared before populating the ARP entry.

How did you verify/test it?
Validate it in the internal setup: arp/test_unknown_mac.py

Signed-off-by: Janetxxx <janet970527@gmail.com>
Signed-off-by: yenlu <yenlu@cisco.com>