[Refine] Use multi-thread to recover chassis (#14951)
Description of PR
Summary:
We're using multi process to reload and recover the chassis.
But the error is not caught well if the issue is happening during recover.
Use multi-thread to obtain better visibility of errors.

Approach
What is the motivation for this PR?
We're using multi process to reload and recover the chassis.
But the error is not caught well if the issue is happening during recover.
Use multi-thread to obtain better visibility of errors.

How did you do it?
Use multi-thread instead of multi-process

How did you verify/test it?
Test it on physical testbed.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>