Fixing the failures in qos-sai on 202405. (#14892)
Description of PR
Summary:
This PR addresses 2 fails in qos-sai runs on cisco-8000 platform. They are:

testQosSaiDscpToPgMapping: The script reads the base counters before starting the actual test loop. However, there is a pretest-packet that is sent before reading the counters. Since there is no delay between sending the pre-test packet and the reading of the base counters, the base counter doesn't include the pre-test packet. In this PR, we are adding a 3 second delay before reading the base counters.
wm_pg_shared_lossy failure: This test is not applicable to T2/multidut/multiasic variants. So the qos.yaml doesn't even have the required key for this test. So we are adding a skip in this PR for this test.

Approach
What is the motivation for this PR?
Test qos sai failures in 202405.

How did you do it?
Pls see the description on how these issues are addressed.

How did you verify/test it?
Ran it in our TB.

Any platform specific information?
Cisco-8000 related.

co-authorized by: jianquanye@microsoft.com