Xfail test script test_snmp_link_local.py (#15082)
What is the motivation for this PR?
There are some issues with running the test script test_snmp_link_local.py on the KVM testbed. In this PR, we have marked the script as 'xfail' and removed it from the onboarding PR checker.

How did you do it?
In this PR, we have marked the script as 'xfail' and removed it from the onboarding PR checker.