added a test to check if same route added on 2 T1 gets propogated to the TOR (#14683)
Description of PR
Summary:
Fixes # (issue)
#10837

Orignal Problem
In multipath scenario when a route is advertised through network command from two neighbors, only one path is chosen. Doing clear bgp * recovers the device from this scenario.

If the same route (be it default or non-default) is advertised from two leaf to a ToR, the leaf initially shows only one leaf's links as next hop. When we do clear bgp *, it is then updated.

B>* 20.0.0.1/32 [20/0] via 192.168.1.2, PortChannel1, weight 1, 00:16:52

               via 192.168.2.2, PortChannel2, weight 1, 00:16:52
               via 192.168.3.2, PortChannel3, weight 1, 00:16:52
               via 192.168.4.2, PortChannel4, weight 1, 00:16:52
After doing clear bgp * in tgn-sonic-n2-t1 I can see this

B>* 20.0.0.1/32 [20/0] via 192.168.1.2, PortChannel1, weight 1, 00:00:02

               via 192.168.2.2, PortChannel2, weight 1, 00:00:02
               via 192.168.3.2, PortChannel3, weight 1, 00:00:02
               via 192.168.4.2, PortChannel4, weight 1, 00:00:02
               via 192.168.5.2, PortChannel5, weight 1, 00:00:02
               via 192.168.6.2, PortChannel6, weight 1, 00:00:02
               via 192.168.8.2, PortChannel8, weight 1, 00:00:02

This issue was fixed in PR sonic-net/sonic-buildimage#17184
However a test for this was missing.
so this test adds a prefix on 2 T1 neighbors using network command and checks if the advertised route is learnt on the T0 DUT

Signed-off-by: yenlu <yenlu@cisco.com>