Merged PR 7794: Modified inventory file to match DUT syseeprom info for Arista 7050CX3 dualtor platforms.
**Context/Motivation**

The syseeprom info in the inventory files for the 4 Arista 7050CX3 DUTs that make up 2 dualtor-aa testbeds do not match what is seen on the devices. This change fixes that issue.

**Fixes MS ADO:** Product Backlog Item 29526116: [SONiC_Nightly][Failed_Case][platform_tests.api.test_chassis.TestChassisApi][test_get_system_eeprom_info][20231110][broadcom][Arista-7050CX3-32S-C32]

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to align inventory file with Device Under Test (DUT) syseeprom information for Arista 7050CX3 dualtor platforms.

#### PR Summary
This pull request updates the inventory file to ensure it matches the syseeprom information for Arista 7050CX3 dualtor platforms.
- Modified `/ansible/strsvc` to update the `0x25` and `0x2E` fields for multiple devices to reflect the correct date and Aboot version.