Fix regression on test_vlan_tc5_untagged_unicast (#14928)
What is the motivation for this PR?
We found error below in running the testcase test_vlan_tc5_untagged_unicast
AssertionError: Received packet that we expected not to receive on device 0, port 43

How did you do it?
Revert the dst_port change in
#12204

How did you verify/test it?
Tested that the test case can pass after this change.

Any platform specific information?

Signed-off-by: yenlu <yenlu@cisco.com>