Fixing reboot-cause for LCs during Kernel Panic testcase (#14986)
Description of PR
In case of Cisco chassis, when supervisor goes for kernel panic based reboot, all LCs are rebooted and the reboot cause is due to Heartbeat loss. Changed the testcase to check the right reboot cause

Summary:
Fixes # (issue)

Approach
What is the motivation for this PR?
In case of Cisco chassis, when supervisor goes for kernel panic based reboot, all LCs are rebooted and the reboot cause is due to Heartbeat loss. Changed the testcase to check the right reboot cause

How did you do it?
How did you verify/test it?
Any platform specific information?
Validated the testcase on Cisco platform

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>