Fix test_lldp_syncd.py (#15072)
test_lldp_syncd.py requires hget_all() funct added to master in PR13574
(https://github.com/sonic-net/sonic-mgmt/pull/13574) but not backported to 202405.

Additionally, the test assumes the value of “lldp_rem_sys_desc” cannot be empty,
which is the case for our switches.

On t2(min), we also require a larger timeout in the wait_until after reboot and to
disable the lognalyzer for the reboot test since there are many error messages during
reboot that are picked up (such as`ERR gbsyncd#syncd: :- updateNotificationsPointers:
pointer for SAI_SWITCH_ATTR_REGISTER_READ is not handled, FIXME!`)

Signed-off-by: yenlu <yenlu@cisco.com>