Add new topology marks to test analysis script (#15177)
What is the motivation for this PR?
In PR #15162, we introduced three new topology marks for certain subtype topologies. To support this change, we are adding these marks to the script that analyzes all test scripts.

How did you do it?
We are adding these new introduced marks to the script that analyzes all test scripts.

Signed-off-by: yenlu <yenlu@cisco.com>