Merged PR 7891: tests-common: Prerequisites for handling PortInUseException for SSHConsoleConn
PR which aligns internal repo with the external repo, allowing for automated tests to pass for the following PR: https://github.com/sonic-net/sonic-mgmt/pull/15174

Accomplishes the following:
1. Adds a console_menu_type for each device under each of the labs in the relevant `_console_links.csv`
2. Adds a `console_username` variable, which is extracted prior to instantiation of the ConsoleHost class (as is done in the external mgmt repo)

----
#### AI description  (iteration 1)
#### PR Classification
Prerequisites for handling `PortInUseException` for `SSHConsoleConn`.

#### PR Summary
This pull request updates the console links configuration to include a new `Console_menu_type` field, which is a prerequisite for handling `PortInUseException` in `SSHConsoleConn`.
- `ansible/files/sonic_bjw2_console_links.csv`: Added `Console_menu_type` field with value `sonic_config` for all entries.