Move the lossy skip condition above. (#15096)
Description of PR
Summary:
The skip for non-single-asic variant for the test:qos/test_qos_sai.py::ttestQosSaiPgSharedWatermark[wm_pg_shared_lossy] was added by #14892 in the wrong location. We need to move this skip before it checks for the wm_pg_shared_lossy key.

Approach
What is the motivation for this PR?
The keyerror addressed in the PR:#14892 is still occuring, since the skip was added after the Key checking code. This PR moves it up above the Keychecking code.

How did you do it?
Moved the skip block above.

How did you verify/test it?
Ran it on my TB:

co-authorized by: jianquanye@microsoft.com