tests-common: Handle PortInUseException for SSHConsoleConn (#15174)
What is the motivation for this PR?
This PR is intended to add more resilience to the dut_console tests, as
if a testbed with a blocked port was selected for a nightly test, it
would cause the dut_console tests to fail during setup. This provides a
method for auto-recovery for these cases.

How did you do it?
By completing the following:

Add connection types for each of the configuration menu types (Digi,
Cisco, and Sonic) to allow for different command sequences to be
accounted for
Add function to clear a DUT's console port
Add retry logic to add resilience to main DUT connection set up
How did you verify/test it?
This process was conducted on 5 testbeds, with at least one of each
config menu type.

Steps conducted during testing
SSH into testbed using the console IP and console port - simulating a
blocked port
In a separate terminal, run any dut_console test individually
During test setup, the connection from step 1 was terminated
successfully
The test then runs successfully
Example test output
Successful port reset (Digi config):
Screenshot 2024-10-25 165601

Successful port reset (Sonic config):
Screenshot 2024-10-25 183646

Sample failure
Simulating config menu type is not defined (causing the port clear
function to exit early), and a more descriptive error message is thrown
regarding why the connection to the DUT is failing:
Screenshot 2024-10-25 153008

Any platform specific information?
N/A - this is a generalised solution which accounts for as many types of
configuration menus as possible.